requirements.txt
openai-whisper
torch
torchaudio
numpy
